refactor: move fillTable function to own module fill_table.js
