+++ nonograms/scripts/game_logic.js
@@ -118,27 +118,6 @@ function getPuzzleNameLevel(isTable) {
   return isTable ? `${puzzleName} - ${puzzleLevel}` : [puzzleName, puzzleLevel];
 }
 
-function fillTable(sign, picArray) {
-  const cellsArray = document.querySelectorAll('.field__cell');
-  picArray.forEach((value, index) => {
-    setTimeout(() => {
-      if (sign.includes('all')) {
-        if (value !== 0) {
-          cellsArray[index].classList.remove('cross');
-          cellsArray[index].classList.add('fill');
-        } else {
-          cellsArray[index].classList.remove('fill');
-          cellsArray[index].classList.add('cross');
-        }
-      }
-      if (value !== 0 && sign.includes('fill')) {
-        cellsArray[index].classList.add('fill');
-      } else if (value !== 0 && sign.includes('cross')) {
-        cellsArray[index].classList.add('cross');
-      }
-    }, (25 / vars.gridSize) * index);
-  });
-}
 
 function setCellStatus(event) {
   const cell = event.target;
